[ADD] Solution for Problem 2
Separate function to calculate the sum of even terms upto limit given.
simple while loop that runs until the term is less than the limit and
a nested if statement adds the term to the sum if it is even. returns
the sum which the main function prints using std::cout