[ADD] Solution for Problem 3
Use `math` library for `sqrt` function. Divide the number out wholly by
integers from 2 to square root of number as the numbers from square root
of n to n have a factor in 2 to square root of n. keep track of every
new factor and return the last one. use a if statement to check if
number is 1 as that means it no longer has any prime factor. break the
while loop at that point. return the last prime factor that divided
number and print it.