[ADD] Solution for Problem 3
Using `math.h` for `sqrt` function. Take a `long int` as input to the
function and return a `int`. divide the number by 2 wholly and then
divide it by each integer from 3 to square root of number skipping the
even numbers in between. for loop and a nested if statement does the
job. exit the loop as soon as number is 1 because then it does not have
a prime factor left. Call the function in the main function and print
the result.