[ADD] Solution for Problem 3
Using `math.h` for the `sqrt` function. define a function that takes
in a `long int` because the number in the problem statement does not
fit in `int`. divide the number tottaly by 2 until its odd then divide
by each number from 3 to square root of number skipping every even
number. exit if number becomes 1. return the last factor that divided
the number. run the function in the main function and print the value