[ADD] Solution for Problem 4
defined a lambda expression for checking palindrome by reversing
the str of int and equating it with the original string representation
of number. for loop that iterated from 999 to 100 decreasing the
variable by one every iteration and a nested for loop that runs from
external for loops counter to 100 similarly. keep the largest
palindrome yet found in a variable and check if the current product is
lesss than the largest palindrome yet found. if it is then break the
inner for loop. if not continue to check if the product is a palindrome
and if it is replace it with the largest palindrome if it is greater
else ignore.
print the largest palindrome in the end.