[ADD] Solution for Problem 4
made a function to check weather a number is *Palindrome* by reversing
it and then checking if its the same as the original number.
for the main function two nested for loops from 999 to 100 and outer
counter to 100 give us all the multiples we need. check if the product
is greater than the largest palindrome that has been found yet and then
if it is check if its a palindrome. if it is a palindrome then replace it
with largest palindrome if it is greater else continue. finally print
the largest palindrome in the end.